fix(hostname): ensure hostname is the FQDN
Previously, the value was compared against `ansible_hostname`,
which only is the FQDN after the first reboot. This lead to the
monitoring not seeing the FQDN as the instance name.

This commit scrapes the hostname using the hostname command,
which is garantueed to return the current running hostname, and
then determines if it needs to be set at runtime.